Move the rest of schema functionality into schema package.
Maybe this commit should be squashed with above lets see

The idea here is to move all the type checking to the schema package
instead of being in yamlmeta and schema.

In this commit we also move the Schema interface to the package that
will need to use it.

Minor changes were made to the tests to better represent the error
messages that we expect to have